requirements.txt
fastapi
uvicorn
python-multipart
litellm
instructor
pandas
diskcache
pydantic
python-dotenv
anthropic
boto3